Movies_MSSQL-main/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==4.0.4
django-cors-headers==3.12.0
django-environ==0.8.1
djangorestframework==3.13.1
idna==3.3
mssql-django==1.1.3
numpy==1.20.2
pip==21.2.4
pyodbc==4.0.32
pytz==2022.1
requests==2.27.1
setuptools==61.2.0
sqlparse==0.4.2
urllib3==1.26.9
wheel==0.37.1
